backend/pyproject.toml
[project]
name = "concurso-coach-ai-backend"
version = "0.1.0"
description = "Backend API for the Concurso Coach AI platform."
authors = [
    {name = "Rodrigo Vieira", email = "rodrigoaugustov@gmail.com"},
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.30",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.2.1",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==3.2.2",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "pydantic[email]>=2.12.3",
    "google-cloud-storage>=3.4.1",
    "redis>=4.2",
    "celery>=5.5.3",
    "google-generativeai>=0.5.0",
    "httpx>=0.28.1",
    "google-genai>=1.45.0",
]

[tool.uv]
# Você pode configurar o comportamento do uv aqui, se necessário.